Remove legacy util function, cleanup names, add _Deprecated flags into filenames
