+++ next/components/molecules/sections/general/ArticlesListSection/BlogPostsByCategory.tsx
@@ -13,9 +13,9 @@ import {
   getBlogPostsQueryKey,
 } from '@/services/meili/fetchers/blogPostsFetcher'
 import { getCategoryColorLocalStyle } from '@/utils/colors'
+import { getNumericLocalDate } from '@/utils/formatDate'
 import { generateImageSizes } from '@/utils/generateImageSizes'
 import { isDefined } from '@/utils/isDefined'
-import { getNumericLocalDate } from '@/utils/local-date'
 import { useRoutePreservedState } from '@/utils/useRoutePreservedState'
 
 const imageSizes = generateImageSizes({ default: '100vw', md: '50vw', lg: '33vw' })
