+++ next/components/molecules/sections/general/ArticlesListSection/InbaArticlesList.tsx
@@ -14,8 +14,8 @@ import {
   inbaArticlesDefaultFilters,
   inbaArticlesFetcher,
 } from '@/services/meili/fetchers/inbaArticlesFetcher'
+import { getNumericLocalDate } from '@/utils/formatDate'
 import { generateImageSizes } from '@/utils/generateImageSizes'
-import { getNumericLocalDate } from '@/utils/local-date'
 import { useRoutePreservedState } from '@/utils/useRoutePreservedState'
 
 const imageSizes = generateImageSizes({ default: '100vw', md: '50vw', lg: '33vw' })
