--- next/utils/local-date.tsx
@@ -1,48 +1,0 @@
-import { convert, LocalDate, ZonedDateTime } from '@js-joda/core'
-import lCapitalize from 'lodash/capitalize'
-import padStart from 'lodash/padStart'
-
-const capitalize = (text: string) => text[0].toUpperCase() + text.slice(1)
-// "Marec 2021"
-export const getLocalDate = (dateString: string) =>
-  dateString
-    ? capitalize(
-        convert(ZonedDateTime.parse(dateString)).toDate().toLocaleDateString('sk', {
-          month: 'long',
-          year: 'numeric',
-        }),
-      )
-    : ''
-
-// "1.3.2021"
-export const getNumericLocalDate = (dateString: string) =>
-  dateString
-    ? convert(ZonedDateTime.parse(dateString))
-        .toDate()
-        .toLocaleDateString('sk', {
-          dateStyle: 'short',
-        })
-        .replace(/ /g, '')
-    : 'Invalid date'
-
-// TODO unify with getNumericLocalDate
-// "01. 03.2021"
-export const formatDate = (dateISOString: string | undefined | null) => {
-  if (!dateISOString) return ''
-  const date = new Date(dateISOString)
-  return date.toLocaleDateString('sk-SK', { month: '2-digit', day: '2-digit', year: 'numeric' })
-}
-
-// "Február"
-export const getLocalMonthName = (month: number | string) => {
-  // TODO !!! get name directly as Month[ix - 1]
-  const validMonth = Math.max(0, Math.round(Number(month))) || 1
-  // eslint-disable-next-line lodash-fp/no-extraneous-args
-  const m = padStart(`${validMonth}`, 2, '0')
-
-  return lCapitalize(
-    convert(LocalDate.parse(`1970-${m}-01`))
-      .toDate()
-      .toLocaleDateString('sk', { month: 'long' }),
-  )
-}
